CPU Management (#2)
Enable the binding of VM's CPUs with host's CPUs

(Depends on https://github.com/PaulRenauld/lsm-perf/pull/1)

This gives the option to the user to use more advanced CPU management with the `-c` flag. When provided with three CPUs x, y and z, x will be used for the I/O and the handling of qemu-system, y will be bound to CPU 0 of the VM, and z to CPU 1. Then, the VM is started with `isolcpus=0` and the workload is run on CPU 0. This means that the workload used a dedicated CPU which should improve the reliability of the measurements. 
The user also needs to start their computer with the kernel option `isolcpus=x,y,z` to unsure that these CPUs are not used for anything else.

For now, [qemu-affinity](https://github.com/zegelin/qemu-affinity) must be cloned and the path to `qemu_affinity.py` must be in `QEMU_AFFINITY_PATH`. This is temporary. The reason is that qemu-affinity needs to be run with `sudo`, which is hard to use with `pip`.